Got all back-end working now (except file edit/file delete functionality), TODO: 1. Front-end. 2. Delete functionality. 3. Edit functionality

+++ src/database/file.go
@@ -1,10 +1,10 @@
 package database
 
 type File struct {
-	ID         uint   `json:"id"`
+	ID         int64  `json:"id"`
 	Location   string `json:"location"`
 	Filename   string `json:"filename"`
 	Public     bool   `json:"public"`
-	Owner      uint   `json:"owner"`
+	Owner      int64  `json:"owner"`
 	ServerName string `json:"serverName"`
 }
